feat: accepting sellet claims
